Books/requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
requests
contextlib
werkzeug.security
functools
os
